Also add cli
